Di::instance renamed to Di::getInstance, new self to new static
