+++ System/Di/DiInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace System\Di;
+
+interface DiInterface {
+    /**
+     * @param string $class
+     * @return object
+     */
+    public function get($class);
+    
+    /**
+     * @return Di
+     */
+    public function reload();
+}
